Add documentation, Move to accessibility folder, Add config
