+++ lib/rubocop/cop/github/accessibility/no_redundant_image_alt.rb
